requirements.txt
uvicorn~=0.24.0.post1
fastapi~=0.104.1
chromadb~=0.4.22
pymongo~=4.6.1
openai~=1.12.0
python-dotenv~=1.0.0